Introduced Flag that tells it to purge/sanitize the gem specification during re-packaging
